+++ TinkoffASDKUI/TinkoffASDKUI/Modules/MainForm/Assembly/IMainFormAssembly.swift
@@ -1,0 +1,12 @@
+//
+//  IMainFormAssembly.swift
+//  TinkoffASDKUI
+//
+//  Created by r.akhmadeev on 19.01.2023.
+//
+
+import UIKit
+
+protocol IMainFormAssembly {
+    func build(stub: MainFormStub) -> UIViewController
+}