+++ TinkoffASDKUI/TinkoffASDKUI/Modules/MainForm/Presenter/IMainFormPresenter.swift
@@ -1,0 +1,13 @@
+//
+//  IMainFormPresenter.swift
+//  TinkoffASDKUI
+//
+//  Created by r.akhmadeev on 19.01.2023.
+//
+
+import Foundation
+
+protocol IMainFormPresenter {
+    func viewDidLoad()
+    func viewWasClosed()
+}