+++ TinkoffASDKUI/TinkoffASDKUI/Modules/MainForm/Router/IMainFormRouter.swift
@@ -1,0 +1,10 @@
+//
+//  IMainFormRouter.swift
+//  TinkoffASDKUI
+//
+//  Created by r.akhmadeev on 19.01.2023.
+//
+
+import Foundation
+
+protocol IMainFormRouter {}