+++ TinkoffASDKUI/TinkoffASDKUI/Modules/MainForm/Router/MainFormRouter.swift
@@ -1,0 +1,12 @@
+//
+//  MainFormRouter.swift
+//  TinkoffASDKUI
+//
+//  Created by r.akhmadeev on 19.01.2023.
+//
+
+import UIKit
+
+final class MainFormRouter: IMainFormRouter {
+    weak var transitionHandler: UIViewController?
+}